Refactor hospital dashboard
