+++ backend/apps/ineedstudent/urls.py
@@ -10,9 +10,8 @@ urlpatterns = [
         StudentSelectionView.as_view(),
         name="list_by_plz",
     ),
-    # path('students_testing/<countrycode>/<plz>/<int:distance>', views.student_list_view, name='student_list_view'),
     path("hospitals/<countrycode>/<plz>", views.hospital_list, name="hospital_list"),
-    path("hospital_map", views.hospital_overview, name="hopsital_map"),
+    path("hospital_map", views.HospitalMapView.as_view(), name="hopsital_map"),
     path("zustimmung", views.zustimmung, name="zustimmung"),
     path("hospital_view/<str:uuid>/", views.hospital_view, name="hospital_view"),
     path("hospital_dashboard", views.HospitalDashboardView.as_view(), name="hospital_dashboard"),
